add finishing touches

+++ src/components/UserPhotos.vue
@@ -44,20 +44,22 @@ export default {
         })
         },
         async deletePhoto(photo) {
-          let image = await db.storage().ref().child(`images/${photo.user}/${photo.name}`);
-          await image.delete().then(() => {
-            }).catch(e => {
-              console.log(e)
-          }) 
-
-          await db.database().ref('images').on('value', snap => {
-            snap.forEach(key => {
-              if(key.val().name === photo.name) {
-                key.ref.remove()
-                this.userPhotos = this.userPhotos.filter(photo => photo.name !== key.val().name)
-              }  
+          if(confirm('Vill du radera denna referens?')) {
+            let image = await db.storage().ref().child(`images/${photo.user}/${photo.name}`);
+            await image.delete().then(() => {
+              }).catch(e => {
+                console.log(e)
+            }) 
+  
+            await db.database().ref('images').on('value', snap => {
+              snap.forEach(key => {
+                if(key.val().name === photo.name) {
+                  key.ref.remove()
+                  this.userPhotos = this.userPhotos.filter(photo => photo.name !== key.val().name)
+                }  
+              })
             })
-          })
+          }
         }
     },
     filters: {
